pyproject.toml
[tool.poetry]
name = "pre-allinone"
version = "0.1.0"
description = ""
authors = ["Magomed Usmanov <mgsmnov@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.1.2"
psycopg = "^3.2.3"
djangorestframework = "^3.15.2"
drf-spectacular = "0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
